baja-cleanup: unified Principal and VeracruzIdentity types
Newly-introduced Principal type was very similar to the existing VeracruzIdentity struct from the veracruz-utils library.  Principal has therefore been removed in favour of a slight generalization of VeracruzIdentity.
